Cleanup uttaksplan-components
