+++ src/uttaksplan/components/uttaksplan/UttaksplanSkjema.tsx
@@ -1,0 +1,69 @@
+import * as React from 'react';
+import EkspanderbartInnhold from 'uttaksplan/elements/ekspanderbartInnhold/EkspanderbartInnhold';
+import { Dekningsgrad, Permisjonsregler } from 'uttaksplan/types';
+import SkjemaDekningsgrad from 'uttaksplan/components/skjema/SkjemaDekningsgrad';
+import SkjemaFordelingFellesperiode from 'uttaksplan/components/skjema/SkjemaFordelingFellesperiode';
+import AktivitetskravInfo from 'uttaksplan/components/content/AktivitetskravInfo';
+
+export interface Props {
+    navnForelder1: string;
+    navnForelder2: string;
+    ukerFellesperiode: number;
+    fellesperiodeukerForelder1: number;
+    dekningsgrad?: Dekningsgrad;
+    permisjonsregler: Permisjonsregler;
+    onChangeDekningsgrad: (dekningsgrad: Dekningsgrad) => void;
+    onChangeFordeling: (uker: number) => void;
+}
+
+class UttaksplanSkjema extends React.Component<Props, {}> {
+    render() {
+        const {
+            navnForelder1,
+            navnForelder2,
+            dekningsgrad,
+            ukerFellesperiode,
+            fellesperiodeukerForelder1,
+            onChangeDekningsgrad,
+            permisjonsregler,
+            onChangeFordeling
+        } = this.props;
+        return (
+            <React.Fragment>
+                <div className="blokk-m">
+                    <SkjemaDekningsgrad
+                        dekningsgrad={dekningsgrad}
+                        antallUkerTotalt80={permisjonsregler.antallUkerTotalt80}
+                        antallUkerTotalt100={
+                            permisjonsregler.antallUkerTotalt100
+                        }
+                        permisjonsregler={permisjonsregler}
+                        onChange={onChangeDekningsgrad}
+                    />
+                </div>
+
+                <EkspanderbartInnhold
+                    animert={false}
+                    erApen={dekningsgrad !== undefined}>
+                    <div className="blokk-s">
+                        <SkjemaFordelingFellesperiode
+                            navnForelder1={navnForelder1}
+                            navnForelder2={navnForelder2}
+                            ukerFellesperiode={ukerFellesperiode}
+                            ukerForelder1={fellesperiodeukerForelder1}
+                            onChange={onChangeFordeling}
+                            introRenderer={() => (
+                                <AktivitetskravInfo
+                                    permisjonsregler={permisjonsregler}
+                                    navnForelder1={navnForelder1}
+                                    navnForelder2={navnForelder2}
+                                />
+                            )}
+                        />
+                    </div>
+                </EkspanderbartInnhold>
+            </React.Fragment>
+        );
+    }
+}
+export default UttaksplanSkjema;
